Range Calculation added

+++ app/src/main/java/com/khs/spcmeasure/entity/Measurement.java
@@ -17,7 +17,7 @@ public class Measurement {
 	private String operator;	// TODO is this required as piece has it?
 	private double value;
     private double range;
-    private Integer cause;
+    private Long cause;
 	private long limitRev;
 	private boolean inControl;
 	private boolean inEngLim;
@@ -27,7 +27,7 @@ public class Measurement {
 
 	// constructor
 	public Measurement(Long id, long pieceId, long prodId, long featId, 
-			Date collectDt, String operator, double value, double range, Integer cause, long limitRev, boolean inControl, boolean inEngLim) {
+			Date collectDt, String operator, double value, double range, Long cause, long limitRev, boolean inControl, boolean inEngLim) {
 		super();
 		this.id = id;
 		this.pieceId = pieceId;
@@ -44,7 +44,7 @@ public class Measurement {
 	}
 
 	public Measurement(long pieceId, long prodId, long featId, 
-			Date collectDt, String operator, double value, double range, Integer cause, long limitRev, boolean inControl, boolean inEngLim) {
+			Date collectDt, String operator, double value, double range, Long cause, long limitRev, boolean inControl, boolean inEngLim) {
 		this(null, pieceId, prodId, featId, collectDt, operator, value, range, cause, limitRev, inControl, inEngLim);
 	}
 
@@ -113,11 +113,11 @@ public class Measurement {
         this.range = range;
     }
 
-    public Integer getCause() {
+    public Long getCause() {
         return cause;
     }
 
-    public void setCause(Integer cause) {
+    public void setCause(Long cause) {
         this.cause = cause;
     }
